Implement Persistence
This commit implements task persistence via an integration with the
sqlite crate. This implementation is simple and directly interfaces with
the DB via SQL queries (as opposed to using an ORM). An alternative for
the simple use-cases within this program would to be use a memory map
which is flushed to disk at certain points, but I found this approach to
be the optimal balance of simplicity and functionality with my
experience.

The main caveat with this approach currently which should be
addressed in future commits is that automated test suites must be run
consecutively due to the tests all accessing the same .db file which causes locking issues.

This commit resolves issue #1.